rename function EnrollRegisteredStudent -> RegisteredStudent
