+++ MainFunction/index.ts
@@ -1,7 +1,7 @@
 import { AzureFunction, Context } from "@azure/functions"
 
 import { XMLParser } from "fast-xml-parser";
-import {isRegistration, enrollRegisteredStudent} from "../EnrollRegisteredStudent";
+import {isRegistration, enrollRegisteredStudent} from "../RegisteredStudent";
 import {isRemoveActivityOccasionApplication, removeActivityEnrollment} from "../RemoveActivityOccasionApplication";
 
 const serviceBusTopicTrigger: AzureFunction = async function(context: Context, message: string): Promise<void> {
@@ -11,9 +11,9 @@ const serviceBusTopicTrigger: AzureFunction = async function(context: Context, m
   const parser = new XMLParser();
   const jsonObj = parser.parse(message);
 
-  const membership = jsonObj?.["ns0:membershipRecord"]?.["ns0:membership"];
+  /* const membership = jsonObj?.["ns0:membershipRecord"]?.["ns0:membership"]; */
 
-  if( isRegistration(membership)){
+  if( isRegistration(jsonObj)){
   /*   context.log("Handle registration message") */
   /*   console.log('::::::::::::;',enrollRegisteredStudent) */
   /*   /1* await enrollRegisteredStudent(context, membership) *1/ */
