+++ RegisteredStudent/index.ts
